requirements.txt
requests==2.27.1
behavex==1.5.3
assertpy==1.1
jsonschema==4.5.1
python-dotenv==0.20.0
pylint==2.13.9
